chore: add internal plop generation for tests
