+++ plop-templates/node-plop-test.js
